mobile_service/mobiles/serializers.py
from rest_framework import serializers
from .models import Mobile


class MobileSerializer(serializers.ModelSerializer):
    class Meta:
        model = Mobile
        fields = [
            "id",
            "name",
            "price",
            "brand",
            "specs",
            "stock",
            "created_at",
            "updated_at",
        ]
        read_only_fields = ["id", "created_at", "updated_at"]
